Restructured wod.forms as folders to match others
