+++ wod/forms/__init__.py
@@ -1,1 +1,1 @@
-from .forms import *
+from . import characters, items, locations
